refactor(mcp-runtime): colocate mobile MCP under AI
