+++ src/ai/mcp/__tests__/McpService.test.ts
