+++ src/ai/mcp/__tests__/resolveAssistantMcpServers.test.ts
