+++ src/ai/mcp/__tests__/schemaBrand.test.ts
