+++ src/ai/mcp/index.ts
@@ -1,0 +1,9 @@
+export type { McpConnectionConfig, McpToolSummary } from './McpService';
+export { McpService } from './McpService';
+export {
+  hasMcpServerWildcardRule,
+  matchesMcpSourceToolRule,
+  withMcpToolRuleAdded,
+  withMcpToolRuleCleared,
+} from './mcpSourcePolicy';
+export { parseFunctionCallToolName } from './mcpToolName';
