+++ src/ai/mcp/mcpResult.ts
