+++ src/ai/mcp/resolveAssistantMcpServers.ts
